+++ flagship/src/main/java/com/abtasty/flagship/decision/IDecisionManager.kt
@@ -1,7 +1,6 @@
 package com.abtasty.flagship.decision
 
 import com.abtasty.flagship.model.Modification
-import com.abtasty.flagship.visitor.VisitorCache
 import com.abtasty.flagship.visitor.VisitorDelegateDTO
 
 interface IDecisionManager {
