+++ flagship/src/main/java/com/abtasty/flagship/visitor/VisitorDelegateDTO.kt
@@ -4,7 +4,7 @@ import org.json.JSONArray
 import org.json.JSONObject
 import java.util.concurrent.ConcurrentLinkedQueue
 
-open class VisitorDelegateDTO(visitorDelegate: VisitorDelegate) {
+open class VisitorDelegateDTO(val visitorDelegate: VisitorDelegate) {
 
     var configManager = visitorDelegate.configManager
     var visitorId = visitorDelegate.visitorId
@@ -15,7 +15,22 @@ open class VisitorDelegateDTO(visitorDelegate: VisitorDelegate) {
     var hasConsented = visitorDelegate.hasConsented
     var isAuthenticated = visitorDelegate.isAuthenticated
     var visitorStrategy = visitorDelegate.getStrategy()
-    var mergedCachedVisitor = visitorDelegate.cachedVisitor
+    var assignmentsHistory = HashMap(visitorDelegate.assignmentsHistory)
+
+    @Suppress("unchecked_cast")
+    override fun toString(): String {
+        val json = JSONObject()
+        json.put("visitorId", visitorId)
+        json.put("anonymousId", if (anonymousId != null) anonymousId else JSONObject.NULL)
+        json.put("isAuthenticated", isAuthenticated)
+        json.put("hasConsented", hasConsented)
+        json.put("context", contextToJson())
+        json.put("modifications", modificationsToJson())
+//        json.put("activatedVariations", activatedVariationToJsonArray(activatedVariations))
+        json.put("activatedVariations", JSONArray(activatedVariations))
+        json.put("assignmentsHistory", JSONObject(assignmentsHistory as Map<Any?, Any?>))
+        return json.toString(2)
+    }
 
     internal fun contextToJson(): JSONObject {
         val contextJson = JSONObject()
@@ -25,27 +40,7 @@ open class VisitorDelegateDTO(visitorDelegate: VisitorDelegate) {
         return contextJson
     }
 
-    override fun toString(): String {
-        val json = JSONObject()
-        json.put("visitorId", visitorId)
-        json.put("anonymousId", if (anonymousId != null) anonymousId else JSONObject.NULL)
-        json.put("isAuthenticated", isAuthenticated)
-        json.put("hasConsented", hasConsented)
-        json.put("context", contextToJson())
-        json.put("modifications", modificationsToJson())
-        json.put("activatedVariations", activatedVariationToJsonArray(activatedVariations))
-        return json.toString(2)
-    }
-
-    private fun activatedVariationToJsonArray(activatedVariations: ConcurrentLinkedQueue<String>) : JSONArray {
-        val array = JSONArray()
-        for (variation in activatedVariations) {
-            array.put(variation)
-        }
-        return array
-    }
-
-    internal fun modificationsToJson(): JSONObject {
+    private fun modificationsToJson(): JSONObject {
         val modificationJson = JSONObject()
         for ((flag, modification) in this.modifications) {
             val value: Any? = modification.value
@@ -53,4 +48,21 @@ open class VisitorDelegateDTO(visitorDelegate: VisitorDelegate) {
         }
         return modificationJson
     }
+
+    fun getVariationGroupAssignment(variationGroupId: String): String? {
+        return assignmentsHistory[variationGroupId]
+    }
+
+    fun addNewAssignmentToHistory(variationGroupId: String?, variationId: String?) {
+        assignmentsHistory[variationGroupId] = variationId
+        visitorDelegate.assignmentsHistory[variationGroupId] = variationId
+    }
+
+//    private fun activatedVariationToJsonArray(activatedVariations: ConcurrentLinkedQueue<String>) : JSONArray {
+//        val array = JSONArray()
+//        for (variation in activatedVariations) {
+//            array.put(variation)
+//        }
+//        return array
+//    }
 }
