+++ flagship/src/test/assets/cache_visitor.json
@@ -18,6 +18,10 @@
       "vip": true,
       "sdk_osName": "Android"
     },
+    "assignmentsHistory": {
+      "brjjpk7734cg0sl5mmmm": "brjjpk7734cg0sl5oooo",
+      "bmsor064jaeg0gm4bbbb": "bmsor064jaeg0gm4dddd"
+    },
     "campaigns": [
       {
         "campaignId": "brjjpk7734cg0sl5llll",
